requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
drf-extra-fields==3.4.1
djoser==2.1.0
pyjwt==2.6.0
pytz==2022.6
sqlparse==0.4.3
gunicorn==20.1.0
psycopg2-binary==2.9.5
Pillow==9.3.0